Add a fragment delegate and remove duplicated code. This allows reusing the fragment code without extending TiFragment or TiDialogFragment
