+++ thirtyinch/src/main/java/net/grandcentrix/thirtyinch/TiDialogFragment.java
@@ -3,22 +3,16 @@ package net.grandcentrix.thirtyinch;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentActivity;
 import android.support.v7.app.AppCompatDialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import net.grandcentrix.thirtyinch.callonmainthread.CallOnMainThreadInterceptor;
-import net.grandcentrix.thirtyinch.distinctuntilchanged.DistinctUntilChangedInterceptor;
 import net.grandcentrix.thirtyinch.internal.InterceptableViewBinder;
-import net.grandcentrix.thirtyinch.internal.PresenterSavior;
-import net.grandcentrix.thirtyinch.internal.PresenterViewBinder;
+import net.grandcentrix.thirtyinch.internal.TiFragmentDelegate;
 import net.grandcentrix.thirtyinch.internal.TiLoggingTagProvider;
 import net.grandcentrix.thirtyinch.internal.TiPresenterProvider;
 import net.grandcentrix.thirtyinch.internal.TiViewProvider;
-import net.grandcentrix.thirtyinch.util.AndroidDeveloperOptions;
-import net.grandcentrix.thirtyinch.util.AnnotationUtil;
 
 import java.util.List;
 
@@ -26,46 +20,33 @@ public abstract class TiDialogFragment<P extends TiPresenter<V>, V extends TiVie
         extends AppCompatDialogFragment implements TiPresenterProvider<P>, TiLoggingTagProvider,
         TiViewProvider<V>, InterceptableViewBinder<V> {
 
-    private static final String SAVED_STATE_PRESENTER_ID = "presenter_id";
-
-    private final String TAG = this.getClass().getSimpleName()
-            + ":" + TiDialogFragment.class.getSimpleName()
-            + "@" + Integer.toHexString(this.hashCode());
-
-    private volatile boolean mActivityStarted = false;
-
-    private P mPresenter;
-
-    private String mPresenterId;
-
-    private PresenterViewBinder<V> mViewBinder = new PresenterViewBinder<>(this);
+    private final TiFragmentDelegate<P, V, ? extends TiDialogFragment> mDelegate = new TiFragmentDelegate<>(this);
 
     @NonNull
     @Override
     public Removable addBindViewInterceptor(@NonNull final BindViewInterceptor interceptor) {
-        return mViewBinder.addBindViewInterceptor(interceptor);
+        return mDelegate.addBindViewInterceptor(interceptor);
     }
 
     @Nullable
     @Override
     public V getInterceptedViewOf(@NonNull final BindViewInterceptor interceptor) {
-        return mViewBinder.getInterceptedViewOf(interceptor);
+        return mDelegate.getInterceptedViewOf(interceptor);
     }
 
     @NonNull
     @Override
-    public List<BindViewInterceptor> getInterceptors(
-            @NonNull final Filter<BindViewInterceptor> predicate) {
-        return mViewBinder.getInterceptors(predicate);
+    public List<BindViewInterceptor> getInterceptors(@NonNull final Filter<BindViewInterceptor> predicate) {
+        return mDelegate.getInterceptors(predicate);
     }
 
     @Override
     public String getLoggingTag() {
-        return TAG;
+        return mDelegate.getLoggingTag();
     }
 
     public P getPresenter() {
-        return mPresenter;
+        return mDelegate.getPresenter();
     }
 
     /**
@@ -74,141 +55,50 @@ public abstract class TiDialogFragment<P extends TiPresenter<V>, V extends TiVie
      */
     @Override
     public void invalidateView() {
-        mViewBinder.invalidateView();
+        mDelegate.invalidateView();
     }
 
     @Override
     public void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-        if (mPresenter == null && savedInstanceState != null) {
-            // recover with Savior
-            // this should always work.
-            final String recoveredPresenterId = savedInstanceState
-                    .getString(SAVED_STATE_PRESENTER_ID);
-            if (recoveredPresenterId != null) {
-                TiLog.v(TAG, "try to recover Presenter with id: " + recoveredPresenterId);
-                //noinspection unchecked
-                mPresenter = (P) PresenterSavior.INSTANCE.recover(recoveredPresenterId);
-                if (mPresenter != null) {
-                    // save recovered presenter with new id. No other instance of this activity,
-                    // holding the presenter before, is now able to remove the reference to
-                    // this presenter from the savior
-                    PresenterSavior.INSTANCE.free(recoveredPresenterId);
-                    mPresenterId = PresenterSavior.INSTANCE.safe(mPresenter);
-                }
-                TiLog.v(TAG, "recovered Presenter " + mPresenter);
-            }
-        }
-
-        if (mPresenter == null) {
-            mPresenter = providePresenter();
-            TiLog.v(TAG, "created Presenter: " + mPresenter);
-            final TiConfiguration config = mPresenter.getConfig();
-            if (config.shouldRetainPresenter() && config.useStaticSaviorToRetain()) {
-                mPresenterId = PresenterSavior.INSTANCE.safe(mPresenter);
-            }
-            mPresenter.create();
-        }
-
-        final TiConfiguration config = mPresenter.getConfig();
-        if (config.isCallOnMainThreadInterceptorEnabled()) {
-            addBindViewInterceptor(new CallOnMainThreadInterceptor());
-        }
-
-        if (config.isDistinctUntilChangedInterceptorEnabled()) {
-            addBindViewInterceptor(new DistinctUntilChangedInterceptor());
-        }
-
-        if (config.shouldRetainPresenter()) {
-            setRetainInstance(true);
-        }
+        mDelegate.onCreate(savedInstanceState);
     }
 
     @Nullable
     @Override
     public View onCreateView(final LayoutInflater inflater, @Nullable final ViewGroup container,
                              @Nullable final Bundle savedInstanceState) {
-        mViewBinder.invalidateView();
+        mDelegate.invalidateView();
         return super.onCreateView(inflater, container, savedInstanceState);
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
-        final FragmentActivity activity = getActivity();
-
-        boolean destroyPresenter = false;
-        if (activity.isFinishing()) {
-            // Probably a backpress and not a configuration change
-            // Activity will not be recreated and finally destroyed, also destroyed the presenter
-            destroyPresenter = true;
-            TiLog.v(TAG, "Activity is finishing, destroying presenter " + mPresenter);
-        }
-
-        final TiConfiguration config = mPresenter.getConfig();
-        if (!destroyPresenter &&
-                !config.shouldRetainPresenter()) {
-            // configuration says the presenter should not be retained, a new presenter instance
-            // will be created and the current presenter should be destroyed
-            destroyPresenter = true;
-            TiLog.v(TAG, "presenter configured as not retaining, destroying " + mPresenter);
-        }
-
-        if (!destroyPresenter &&
-                !config.useStaticSaviorToRetain()
-                && AndroidDeveloperOptions.isDontKeepActivitiesEnabled(getActivity())) {
-            // configuration says the PresenterSavior should not be used. Retaining the presenter
-            // relays on the Activity nonConfigurationInstance which is always null when
-            // "don't keep activities" is enabled.
-            // a new presenter instance will be created and the current presenter should be destroyed
-            destroyPresenter = true;
-            TiLog.v(TAG, "the PresenterSavior is disabled and \"don\'t keep activities\" is "
-                    + "activated. The presenter can't be retained. Destroying " + mPresenter);
-        }
-
-        if (destroyPresenter) {
-            mPresenter.destroy();
-            PresenterSavior.INSTANCE.free(mPresenterId);
-        } else {
-            TiLog.v(TAG, "not destroying " + mPresenter
-                    + " which will be reused by the next Activity instance, recreating...");
-        }
+        mDelegate.onDestroy();
     }
 
     @Override
     public void onDestroyView() {
-        mPresenter.detachView();
+        mDelegate.onDestroyView();
         super.onDestroyView();
     }
 
     @Override
     public void onSaveInstanceState(final Bundle outState) {
         super.onSaveInstanceState(outState);
-        outState.putString(SAVED_STATE_PRESENTER_ID, mPresenterId);
+        mDelegate.onSaveInstanceState(outState);
     }
 
     @Override
     public void onStart() {
         super.onStart();
-        mActivityStarted = true;
-
-        if (isUiPossible()) {
-            getActivity().getWindow().getDecorView().post(new Runnable() {
-                @Override
-                public void run() {
-                    if (isUiPossible() && mActivityStarted) {
-                        mViewBinder.bindView(mPresenter, TiDialogFragment.this);
-                    }
-                }
-            });
-        }
+        mDelegate.onStart();
     }
 
     @Override
     public void onStop() {
-        mActivityStarted = false;
-        mPresenter.detachView();
+        mDelegate.onStop();
         super.onStop();
     }
 
@@ -220,40 +110,11 @@ public abstract class TiDialogFragment<P extends TiPresenter<V>, V extends TiVie
      */
     @NonNull
     public V provideView() {
-
-        final Class<?> foundViewInterface = AnnotationUtil
-                .getInterfaceOfClassExtendingGivenInterface(this.getClass(), TiView.class);
-
-        if (foundViewInterface == null) {
-            throw new IllegalArgumentException(
-                    "This Fragment doesn't implement a TiView interface. "
-                            + "This is the default behaviour. Override provideView() to explicitly change this.");
-        } else {
-            if (foundViewInterface.getSimpleName().equals("TiView")) {
-                throw new IllegalArgumentException(
-                        "extending TiView doesn't make sense, it's an empty interface."
-                                + " This is the default behaviour. Override provideView() to explicitly change this.");
-            } else {
-                // assume that the fragment itself is the view and implements the TiView interface
-                //noinspection unchecked
-                return (V) this;
-            }
-        }
+        return mDelegate.provideView();
     }
 
     @Override
     public String toString() {
-        String presenter = getPresenter() == null ? "null" :
-                getPresenter().getClass().getSimpleName()
-                        + "@" + Integer.toHexString(getPresenter().hashCode());
-
-        return getClass().getSimpleName()
-                + ":" + TiDialogFragment.class.getSimpleName()
-                + "@" + Integer.toHexString(hashCode())
-                + "{presenter=" + presenter + "}";
-    }
-
-    private boolean isUiPossible() {
-        return isAdded() && !isDetached();
+        return mDelegate.fragmentToString();
     }
 }
